requirements.txt
opentelemetry-sdk==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
